# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/videos": {
      "post": {
        "summary": "创建视频 ",
        "deprecated": false,
        "description": "OpenAI 兼容的视频生成接口。\n\n参考文档: https://platform.openai.com/docs/api-reference/videos/create\n",
        "operationId": "createVideo",
        "tags": [
          "视频生成/Sora兼容格式"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "description": "模型名称",
                    "example": "sora-2",
                    "type": "string"
                  },
                  "prompt": {
                    "description": "提示词",
                    "example": "cute cat dance",
                    "type": "string"
                  },
                  "seconds": {
                    "description": "生成秒数",
                    "example": "8",
                    "type": "string"
                  },
                  "duration": {
                    "description": "生成时长（秒），兼容非 OpenAI 视频协议",
                    "example": 8,
                    "type": "number",
                    "minimum": 0,
                    "maximum": 3600
                  },
                  "size": {
                    "description": "输出尺寸，例如 1280x720",
                    "example": "1280x720",
                    "type": "string"
                  },
                  "resolution": {
                    "description": "输出分辨率，例如 720p/1080p",
                    "example": "720p",
                    "type": "string"
                  },
                  "aspect_ratio": {
                    "description": "画面比例，例如 16:9",
                    "example": "16:9",
                    "type": "string"
                  },
                  "quality": {
                    "description": "质量档位（由上游模型决定）",
                    "example": "high",
                    "type": "string"
                  },
                  "width": {
                    "description": "输出宽度",
                    "example": 1280,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 16384
                  },
                  "height": {
                    "description": "输出高度",
                    "example": 720,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 16384
                  },
                  "fps": {
                    "description": "帧率",
                    "example": 24,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 240
                  },
                  "n": {
                    "description": "生成数量",
                    "example": 1,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 16
                  },
                  "seed": {
                    "description": "随机种子",
                    "example": 42,
                    "type": "integer",
                    "format": "int64"
                  },
                  "reference_images": {
                    "description": "参考图片 URL 列表",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "first_frame_image": {
                    "description": "首帧图片 URL 或 Base64",
                    "type": "string"
                  },
                  "last_frame_image": {
                    "description": "尾帧图片 URL 或 Base64",
                    "type": "string"
                  },
                  "images": {
                    "description": "多图输入（可重复提交该字段）",
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "response_format": {
                    "description": "结果格式",
                    "example": "url",
                    "type": "string"
                  },
                  "user": {
                    "description": "调用方用户标识",
                    "type": "string"
                  },
                  "input_reference": {
                    "format": "binary",
                    "type": "string",
                    "description": "参考图片文件",
                    "example": ""
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建视频任务",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "视频 ID"
                    },
                    "object": {
                      "type": "string",
                      "description": "对象类型"
                    },
                    "model": {
                      "type": "string",
                      "description": "使用的模型"
                    },
                    "status": {
                      "type": "string",
                      "description": "任务状态"
                    },
                    "progress": {
                      "type": "integer",
                      "description": "进度百分比"
                    },
                    "created_at": {
                      "type": "integer",
                      "description": "创建时间戳"
                    },
                    "seconds": {
                      "type": "string",
                      "description": "视频时长"
                    },
                    "completed_at": {
                      "type": "integer",
                      "description": "完成时间戳"
                    },
                    "expires_at": {
                      "type": "integer",
                      "description": "过期时间戳"
                    },
                    "size": {
                      "type": "string",
                      "description": "视频尺寸"
                    },
                    "error": {
                      "$ref": "#/components/schemas/OpenAIVideoError"
                    },
                    "metadata": {
                      "type": "object",
                      "description": "额外元数据",
                      "additionalProperties": true,
                      "properties": {}
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "model",
                    "status",
                    "progress",
                    "created_at",
                    "seconds"
                  ]
                },
                "example": {
                  "id": "sora-2-123456",
                  "object": "video",
                  "model": "sora-2",
                  "status": "queued",
                  "progress": 0,
                  "created_at": 1764347090922,
                  "seconds": "8"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/videos/{task_id}": {
      "get": {
        "summary": "获取视频任务状态 ",
        "deprecated": false,
        "description": "OpenAI 兼容的视频任务状态查询接口。\n\n返回视频任务的详细状态信息。\n",
        "operationId": "getVideo",
        "tags": [
          "视频生成/Sora兼容格式"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "视频任务 ID",
            "required": true,
            "example": "sora-2-123456",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取视频任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "progress": {
                      "type": "integer"
                    },
                    "created_at": {
                      "type": "integer"
                    },
                    "seconds": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "model",
                    "status",
                    "progress",
                    "created_at",
                    "seconds"
                  ]
                },
                "example": {
                  "id": "sora-2-123456",
                  "object": "video",
                  "model": "sora-2",
                  "status": "queued",
                  "progress": 0,
                  "created_at": 1764347090922,
                  "seconds": "8"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "取消视频任务",
        "description": "取消尚未完成的视频任务并退还未结算额度。上游取消接口由渠道适配器尽力调用；网关本地任务状态以原子方式变更为 failed。",
        "operationId": "cancelVideo",
        "tags": [
          "视频生成/Sora兼容格式"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "视频任务 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "任务已取消",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIVideo"
                }
              }
            }
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "任务已完成或已被其他请求取消",
            "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/video/generations": {
      "post": {
        "summary": "创建视频生成任务",
        "deprecated": false,
        "description": "提交视频生成任务，支持文生视频和图生视频。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员在月球上漫步",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/video/generations/{task_id}": {
      "get": {
        "summary": "获取视频生成任务状态",
        "deprecated": false,
        "description": "查询视频生成任务的状态和结果。\n\n任务状态：\n- `queued`: 排队中\n- `in_progress`: 生成中\n- `completed`: 已完成\n- `failed`: 失败\n",
        "operationId": "getVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "任务 ID",
            "required": true,
            "example": "abcd1234efgh",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "VideoRequest": {
        "type": "object",
        "description": "视频生成请求",
        "properties": {
          "model": {
            "type": "string",
            "description": "模型/风格 ID",
            "example": "kling-v1"
          },
          "prompt": {
            "type": "string",
            "description": "文本描述提示词",
            "example": "宇航员站起身走了"
          },
          "image": {
            "type": "string",
            "description": "图片输入 (URL 或 Base64)",
            "example": "https://example.com/image.jpg"
          },
          "images": {
            "type": "array",
            "description": "多图/首尾帧输入（按顺序传递）",
            "items": {
              "type": "string"
            }
          },
          "input_reference": {
            "type": "string",
            "description": "OpenAI 视频协议的参考图片 URL 或 Base64"
          },
          "reference_images": {
            "type": "array",
            "description": "参考图片 URL 列表",
            "items": {
              "type": "string"
            }
          },
          "first_frame_image": {
            "type": "string",
            "description": "首帧图片 URL 或 Base64"
          },
          "last_frame_image": {
            "type": "string",
            "description": "尾帧图片 URL 或 Base64"
          },
          "duration": {
            "type": "number",
            "description": "视频时长（秒）",
            "example": 5
          },
          "width": {
            "type": "integer",
            "description": "视频宽度",
            "example": 1280,
            "minimum": 0,
            "maximum": 16384
          },
          "height": {
            "type": "integer",
            "description": "视频高度",
            "example": 720,
            "minimum": 0,
            "maximum": 16384
          },
          "resolution": {
            "type": "string",
            "description": "输出分辨率（例如 720p、1080p）",
            "example": "720p"
          },
          "aspect_ratio": {
            "type": "string",
            "description": "画面比例",
            "example": "16:9"
          },
          "quality": {
            "type": "string",
            "description": "质量档位"
          },
          "fps": {
            "type": "integer",
            "description": "视频帧率",
            "example": 30
          },
          "seed": {
            "type": "integer",
            "description": "随机种子",
            "example": 20231234
          },
          "n": {
            "type": "integer",
            "description": "生成视频数量",
            "example": 1
          },
          "response_format": {
            "type": "string",
            "description": "响应格式",
            "example": "url"
          },
          "user": {
            "type": "string",
            "description": "用户标识",
            "example": "user-1234"
          },
          "metadata": {
            "type": "object",
            "description": "扩展参数 (如 negative_prompt, style, quality_level 等)",
            "additionalProperties": true,
            "properties": {}
          }
        }
      },
      "VideoResponse": {
        "type": "object",
        "description": "视频生成任务提交响应",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任务 ID",
            "example": "abcd1234efgh"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "example": "queued"
          }
        }
      },
...
      "OpenAIVideo": {
        "type": "object",
        "description": "OpenAI 兼容的视频对象",
        "properties": {
          "id": {
            "type": "string",
            "description": "视频 ID",
            "example": "video-abc123"
          },
          "task_id": {
            "type": "string",
            "description": "任务 ID (兼容旧接口)",
            "deprecated": true
          },
          "object": {
            "type": "string",
            "description": "对象类型",
            "example": "video"
          },
          "model": {
            "type": "string",
            "description": "使用的模型",
            "example": "sora"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "enum": [
              "queued",
              "in_progress",
              "completed",
              "failed"
            ],
            "example": "completed"
          },
          "progress": {
            "type": "integer",
            "description": "进度百分比",
            "example": 100
          },
          "created_at": {
            "type": "integer",
            "description": "创建时间戳"
          },
          "completed_at": {
            "type": "integer",
            "description": "完成时间戳"
          },
          "expires_at": {
            "type": "integer",
            "description": "过期时间戳"
          },
          "seconds": {
            "type": "string",
            "description": "视频时长"
          },
          "size": {
            "type": "string",
            "description": "视频尺寸"
          },
          "remixed_from_video_id": {
            "type": "string",
            "description": "源视频 ID（如果是基于其他视频生成）"
          },
          "error": {
            "$ref": "#/components/schemas/OpenAIVideoError"
          },
          "metadata": {
            "type": "object",
            "description": "额外元数据",
            "additionalProperties": true,
            "properties": {}
          }
        }
      },
      "OpenAIVideoError": {
        "type": "object",
        "description": "OpenAI 视频错误信息",
        "properties": {
          "message": {
            "type": "string",
            "description": "错误信息"
          },
          "code": {
            "type": "string",
            "description": "错误码"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}